LPS-171387 Centralize all release checks in ReleaseManagerImpl
