+++ modules/apps/portal/portal-upgrade-impl/src/main/java/com/liferay/portal/upgrade/internal/release/osgi/commands/ReleaseManagerOSGiCommands.java
@@ -17,26 +17,16 @@ package com.liferay.portal.upgrade.internal.release.osgi.commands;
 import com.liferay.gogo.shell.logging.TeeLoggingUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.kernel.upgrade.UpgradeStep;
-import com.liferay.portal.kernel.upgrade.util.UpgradeProcessUtil;
-import com.liferay.portal.kernel.version.Version;
-import com.liferay.portal.upgrade.PortalUpgradeProcess;
 import com.liferay.portal.upgrade.internal.executor.UpgradeExecutor;
 import com.liferay.portal.upgrade.internal.graph.ReleaseGraphManager;
 import com.liferay.portal.upgrade.internal.registry.UpgradeInfo;
 import com.liferay.portal.upgrade.internal.release.ReleaseManagerImpl;
 
-import java.sql.Connection;
-import java.sql.SQLException;
-
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.SortedMap;
 
 import org.apache.felix.service.command.Descriptor;
 
@@ -59,12 +49,12 @@ public class ReleaseManagerOSGiCommands {
 
 	@Descriptor("List pending upgrades")
 	public String check() {
-		return _check(false);
+		return _releaseManagerImpl.check(false);
 	}
 
 	@Descriptor("List pending upgrade processes and their upgrade steps")
 	public String checkAll() {
-		return _check(true);
+		return _releaseManagerImpl.check(true);
 	}
 
 	@Descriptor("Execute upgrade for a specific module")
@@ -223,176 +213,6 @@ public class ReleaseManagerOSGiCommands {
 				}
 			}
 		}
-	}
-
-	private String _check(boolean showUpgradeSteps) {
-		StringBundler sb = new StringBundler(3);
-
-		sb.append(_checkPortal(showUpgradeSteps));
-
-		if (sb.length() > 0) {
-			sb.append(StringPool.NEW_LINE);
-		}
-
-		sb.append(_checkModules(showUpgradeSteps));
-
-		return sb.toString();
-	}
-
-	private String _checkModules(boolean showUpgradeSteps) {
-		StringBundler sb = new StringBundler();
-
-		Set<String> bundleSymbolicNames =
-			_releaseManagerImpl.getBundleSymbolicNames();
-
-		for (String bundleSymbolicName : bundleSymbolicNames) {
-			String schemaVersionString =
-				_releaseManagerImpl.getSchemaVersionString(bundleSymbolicName);
-
-			ReleaseGraphManager releaseGraphManager = new ReleaseGraphManager(
-				_releaseManagerImpl.getUpgradeInfos(bundleSymbolicName));
-
-			List<List<UpgradeInfo>> upgradeInfosList =
-				releaseGraphManager.getUpgradeInfosList(schemaVersionString);
-
-			int size = upgradeInfosList.size();
-
-			if (size > 1) {
-				sb.append("There are ");
-				sb.append(size);
-				sb.append(" possible end nodes for ");
-				sb.append(schemaVersionString);
-				sb.append(StringPool.NEW_LINE);
-			}
-
-			if (size == 0) {
-				continue;
-			}
-
-			List<UpgradeInfo> upgradeInfos = upgradeInfosList.get(0);
-
-			UpgradeInfo lastUpgradeInfo = upgradeInfos.get(
-				upgradeInfos.size() - 1);
-
-			sb.append(
-				_getModulePendingUpgradeMessage(
-					bundleSymbolicName, schemaVersionString,
-					lastUpgradeInfo.getToSchemaVersionString()));
-
-			if (showUpgradeSteps) {
-				sb.append(StringPool.COLON);
-
-				for (UpgradeInfo upgradeInfo : upgradeInfos) {
-					UpgradeStep upgradeStep = upgradeInfo.getUpgradeStep();
-
-					sb.append(StringPool.NEW_LINE);
-					sb.append(StringPool.TAB);
-					sb.append(
-						_getPendingUpgradeProcessMessage(
-							upgradeStep.getClass(),
-							upgradeInfo.getFromSchemaVersionString(),
-							upgradeInfo.getToSchemaVersionString()));
-				}
-			}
-
-			sb.append(StringPool.NEW_LINE);
-		}
-
-		return sb.toString();
-	}
-
-	private String _checkPortal(boolean showUpgradeSteps) {
-		try (Connection connection = DataAccess.getConnection()) {
-			Version currentSchemaVersion =
-				PortalUpgradeProcess.getCurrentSchemaVersion(connection);
-
-			SortedMap<Version, UpgradeProcess> pendingUpgradeProcesses =
-				PortalUpgradeProcess.getPendingUpgradeProcesses(
-					currentSchemaVersion);
-
-			if (!pendingUpgradeProcesses.isEmpty()) {
-				Version latestSchemaVersion =
-					PortalUpgradeProcess.getLatestSchemaVersion();
-
-				StringBundler sb = new StringBundler();
-
-				sb.append(
-					_getModulePendingUpgradeMessage(
-						"Portal", currentSchemaVersion.toString(),
-						latestSchemaVersion.toString()));
-
-				sb.append(" (requires upgrade tool or auto upgrade)");
-
-				if (showUpgradeSteps) {
-					sb.append(StringPool.COLON);
-
-					for (SortedMap.Entry<Version, UpgradeProcess> entry :
-							pendingUpgradeProcesses.entrySet()) {
-
-						sb.append(StringPool.NEW_LINE);
-						sb.append(StringPool.TAB);
-
-						UpgradeProcess upgradeProcess = entry.getValue();
-						Version version = entry.getKey();
-
-						sb.append(
-							_getPendingUpgradeProcessMessage(
-								upgradeProcess.getClass(),
-								currentSchemaVersion.toString(),
-								version.toString()));
-
-						sb.append(StringPool.NEW_LINE);
-
-						currentSchemaVersion = version;
-					}
-				}
-
-				return sb.toString();
-			}
-		}
-		catch (SQLException sqlException) {
-			if (_log.isWarnEnabled()) {
-				_log.warn(
-					"Unable to get pending upgrade information for the portal",
-					sqlException);
-			}
-		}
-
-		return StringPool.BLANK;
-	}
-
-	private String _getModulePendingUpgradeMessage(
-		String moduleName, String currentSchemaVersion,
-		String finalSchemaVersion) {
-
-		return StringBundler.concat(
-			"There are upgrade processes available for ", moduleName, " from ",
-			currentSchemaVersion, " to ", finalSchemaVersion);
-	}
-
-	private String _getPendingUpgradeProcessMessage(
-		Class<?> upgradeClass, String fromSchemaVersion,
-		String toSchemaVersion) {
-
-		StringBundler sb = new StringBundler(6);
-
-		String toMessage = toSchemaVersion;
-
-		if (UpgradeProcessUtil.isRequiredSchemaVersion(
-				Version.parseVersion(fromSchemaVersion),
-				Version.parseVersion(toSchemaVersion))) {
-
-			toMessage += " (REQUIRED)";
-		}
-
-		sb.append(fromSchemaVersion);
-		sb.append(" to ");
-		sb.append(toMessage);
-		sb.append(StringPool.COLON);
-		sb.append(StringPool.SPACE);
-		sb.append(upgradeClass.getName());
-
-		return sb.toString();
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
